SPRNGCORE-5: Upgrade to spring 3.3.7.
This implements a custom UUID generator based because the old generator was deprecated and is now no longer available in Spring.

The new generator is called `FolioUUIDGenerator` (replacing `CustomUUIDGenerator`) to avoid possible name conflicts and confusion with upstream generators.

A custom `JacksonConfiguration` is provided to help ensure proper building of the JSON due to behavioral changes following the upgrade.
There may need to be future work on the mapping settings as the options chosen are not yet well tested.